Work with the "browser" object instead of "chrome".
Installed the "webextension-polyfill" library to have support for "browser" in chromium-based browsers.
Changed the code to work with "browser" which is promise-based instead of callbacks-based.
Simplified the code to load and save the Jira-RTL settings inside the browser storage.

Inside main.js, instead of listening to runtime messages, listening to changes in the browser storage instead.
This simplifies the code in popup.js to just update the storage without transmitting any messages.

Acting on changed settings based on the changed properties only, by comparing them to previous values.