Improved css of cards titles in Sprint or Kanban view.
Jira has changed the card title container to behave differently after hovering on it (the DOM layout is changed).
Changed it so after hovering the direction of the parent container will be back to LTR, and the direction of the text itself will be RTL.
Improved position of the edit icon of the card title, in RTL layout - sometimes it can still show outside the card box.